NAS-100746 Move license button to top section
