+++ src/app/pages/system/support/sys-info/tn-sys-info.component.html
@@ -13,6 +13,7 @@
         <div class="tn-sys-info-line"><h4>Model:</h4> {{model}}</div>
         <div class="tn-sys-info-line"><h4>System Serial:</h4> {{sys_serial}}</div>
         <div class="tn-sys-info-line"><h4>Additional Hardware:</h4> {{add_hardware}}</div>
+        <div><button mat-flat-button id="update-license-btn" (click)="updateLicense()">{{"UPDATE LICENSE" | translate}}</button></div>
     </div>
   </div>
   <div id="tn-sys-info" class="container" fxLayout="row">
